Adds support for round-trippng through the tokenizer as the 1st step in supporting some basic refactorings.  So far this is all about successful tokenization - nothing for malformed tokens.  Adds tests which round-trip the standard library for Python 2.4 - Python 3.2.
Adds a new VerbatimImage to each token and we create verbatim images which capture the original text where necessary when parsing in verbatim mode.  For new line tokens we have a newline token for each different possible new line image (and again for the NL tokens).

We also now track the white space proceeding each token and make it available in verbatim mode.

We used to have 2 paths for reading new lines, unifies on one and optimizes the indentation consistency checks so they don't need to allocate a string builder.  This simplifies things so we don't need to update the 2 different code paths for tracking the proceeding white space.  There was also at least one difference in these functions which resulted in different spans being reported for errors.  We now unify on behavior of the version which didn't check for inconsistency.

Also:
	removes the ProjectSmuggler public API surface and instead goes through EnvDTE to get back to our project.
	Fixes a 3.2 incompatibility where it's now legal to delete a closure variable - also enables parsing the 3.2 standard library which was missing from the parser tests