Implements the ?Extract Method? refactoring.
To support this we need to have name binding (really reference binding) re-enabled so we understand where variables go to.  This is moved outside of the nodes (so we?re not mutating nodes to bind names) and changes it so that a PythonReference instance exists for each unique location in the source code where we refer to variables.  Having the references be unique makes it much easier to process the references when we?re trying to figure out what was referenced in the extracted code vs. what was referenced in the surrounding code.

We also pull the FlowChecker out of the analysis DLL (where it?s completely not needed ? extract method needs flow checking, but otherwise we don?t need it because we don?t compile).    The FlowChecker is completely updated to store all state outside of the AST nodes.

Tweaked some of our round-tripping behavior:
	Fixed an issue where we would not round trip ?def f(a, (b? because we would add an extra closing paren for b.
	SuiteStatement no longer gets the indentation white space for the next token, instead that goes onto the statement we will parse.
	Tweaked how SuiteStatement is marked
	Adde some APIs onto various AST nodes for exposing some of the white space and setting white space attributes

Makes PythonVariable/PythonReference public so that we can inspect the bound trees
Makes non-recursive walker public so we can use it


Fixes an issue where comments weren?t being classified properly because we were passing in the wrong options to the tokenizer
Adds some new mock classes (MockTextOptions, MockTextSelction, MockTextEdit) so we can run refactoring tests w/o visual studio

At a high level extract method works as follows:
	First we walk the AST to find the a node which encloses the entire selection (minus any white space on the edges of the selection)
	If we had to expand the selection past selected nodes we ask the user if they?re cool with that
	We then run a bunch of walkers which check if we?re performing a refactoring we cannot handle:
		Break/continue w/o the enclosing loop
		Extracting statements from a class definition
		Extracting a node which is unextractable (e.g a parameter name)
		Extracting something which contains a yield
		Extracting something which contains an import *
		Extracting something which contains a return but where not all paths are terminated (by either a return or a raise)
	If the extraction is valid then we need to figure out the variables which are used:
First we walk the extracted body to find everything which is read/assigned to
Then we run the flow checker on the body to discover any variables which are accessed without being assigned
Then we walk the entire AST to find the variables which are assigned to before the extracted method or read after the extracted method

Once we?ve done that we know all the appropriate variables.   Finally we display the UI to the user which lets them select the target scope and based upon that scope what variables should remain as closure variables.

	Once the user has OKed the refactoring we do the re-write.  We create a new FunctionDefinition which has the extracted body.  We setup the new function so that it first has any self or cls parameter (if we?re extracting to/from a class), then we add all of the parameters which are consumed (in alphabetical order), and finally we add the variables the user does not want closed over.  	If the function returns output variables then we add a return statement to the end of the function definition body.

	We then take the newly generated function and turn it back into code by calling ToCodeString.  The function will not be at an incorrect (and mismatched) indentation level.  We walk the body (now the def & decorators) to discover the current indentation level.  We then walk each line removing the indentation and adding in the correct level of indentation.

	Finally we need to build the call to the function which we do by just building up the appropriate string.  If we have output variables we add the assignment to them, otherwise if the extracted statement returned we add the return statement (you can?t have both).  If we?re extracting a method into a class then we add the appropriate ?self.? or ?C.? depending on if the method is an instance method ,class method, or a static method.    Finally we add all of the input parameters into the call.

	The method extractor then deletes the previous selected code, inserts the call site, and inserts the function definition at the last valid statement before where we?re extracting it from (so it?s at the correct location if we?re closing over variables).