424 Generating IronPython completion database w/ Scipy/numpy for .NET installed fails 	Impact: IronPython users w/ numpy/scipy installed will not get completions for some portions of the Python standard library 	Workaround: Users would need to update one of the Python scripts to avoid generating a bad (empty) database file
When scraping built-in modules for their members we can come up with a module name which isn?t really importable for one reason or another.  When this happens we still write out a intellisense database but it contains the pickle for None ? which we turn back into a null reference when we load the database.  We then crash trying to dereference the null pointer that we think should be a dictionary.

The fix is to not write out the database if we get back None.