434 Cannot paste part of a line into the repl 	Take 2 after John?s feedback from the previous code review.  We now always combine the current text, split it, and then submit all but the last input.  When we?re ready for the next input we?ll set the last input to the current input.
523 When the correct encoding is declared on the #coding line we should automatically use that encoding to open a file
513 the problem with python coding
	The method by which we were providing the encoding to Visual Studio was being ignored.  Switch to using MEF API which is respected.

Also: When Tools->Options->Projects and Solutions->Save new projects when created is unchecked, you create a new project, save it, and then attempt to rename a file you get an exception.  This is because we still have cached the old location for the project and the rename attempts to rename relative to the old dir.  Now we clear the old dir when the save happens.

Also when you then attempt to re-open the solution the files aren?t actually in the correct locations.

Also fixes our #1 watson bug which is caused by our bug (we have another Watson bug which causes the most crashes ? but it?s caused by the debugger canvas).  When editing the interpreters you can click on the ??? button even when the text box is disabled and there are no interpreters selected.  This crashes Visual Studio.