Adds section on Just my Code to the spec.
The Visual Studio ?Just my Code? feature enables finer grain control of when to break on an exception.


Enabling or disabling the ?Just my Code? setting in Tools->Options->Debugging will alter how the Exceptions dialog lets the user configure handling of exceptions. When this option is enabled the user will have the option of specifying the behavior of exceptions when they are thrown and when they?re user-unhandled:


When the option is disabled the user can only specify whether to break when the exception is thrown:


Breaking when Thrown
When the user has selected to break when an exception is thrown the debugger will break in whenever that exception is raised.
Independently of Just my Code users can enable breaking of SystemExit exceptions which have an exit code of 0. When the user has not enabled this option SystemExit exceptions with a zero exit code will be ignored.
Break when User-unhandled
When the user has selected to break when an exception is user-unhandled the debugger will break only if we cannot find a handler for the exception within the users? code.
To determine if the exception is user-unhandled we currently search the current call stack to see if any of the calling frames are surrounded with a try/except statement.
If a try/except statement is encountered we check and see if it either catches all exceptions or if it is simply catching an exception by name or by a qualified member expression (for example ?Exception? or ?socket.Error?). If the expression for determining the type of exception that the except block handles is more complicated it will be ignored to avoid introducing unexpected side effects while debugging.
If the type of the exception being raised matches the type of exception the except block handles then the exception will not be reported to the user.
After searching all of the stack frames if a handler has not been found then the debugger will break in and report the unhandled exception to the user.
Exception Dialog
When the exception is reported to the user they are informed whether we are breaking for an unhandled or handled exception.

An exception which is thrown and handled is reported to the user with a dialog like this:


An unhandled exception is reported to the user with a dialog like this: