Fix regression to refactor rename caused by change 25888 on 10/24/2011.
The issue here is that we aren?t updating the analyze version of a located variable def (function defs and class defs being the particular problem).  When we open the project we?ll parse a file from the project system and re-parse it when it gets opened in an editor buffer.  This causes the versions  to be out of sync and when we go and do the rename we don?t provide the definition location ? which causes us not to rename the definition, instead we only rename the references.  Our test cases all pass though as they only parse once.

Fix the issue where we?re not updating the analyze version, update the test cases so they attempt the test cases a 2nd time w/ a re-parsed buffer.

This appears to expose an additional issue w/ with refactorings that I?ll hold off on until after Beta.