Fix an issue w/ generators where extract method throws a null ref exception that I ran into.
On 4/12/2011 I added a check to PythonNameBinder so that we wouldn?t name bind generators ? this was intended for generator expressions, not function definitions.  Before these were effectively the same as we didn?t yet have GeneratorExpression inheriting from Comprehension (and there was no way to distinguish the two).  But along the way to 1.0 GeneratorExpression became a normal Comprehension (there?s no FunctionDefinition involved at all) so this check is no longer necessary.

We are also missing a check to see if a FuncDef is a lambda when doing extract method ? this causes extract method to blow up when there?s a lambda present.

Adds test cases for both, along with a test case for extracting the default value of a parameter.