Fix an MPFProj bug 11618 I ran into while trying to repro bug 626 When copying a folder which contains multiple children MPFProj has a bug where they continue walking the same node repeatedly instead of moving onto the next sibling node.  Now we properly move onto the next sibling. Adds a new test case as BasicProjectTests.CopyFolderWithMultipleItems
Also fixes a couple of project tests which are currently failing:
	RenameProjectTest/ProjectRenameAndDeleteItem ? when fixing bug 497 we started allowing filenames w/ leading dots
	ProjectAddItem ? we added the references node in 1.1 which throws off a couple of assertions
	TestRelativePaths ? we added linked nodes in 1.1 which broke this test

Also fixes a spot where we prompt even if we?re in an automation function
And fixes raising the exited event from debugging.  We were hooking the event, but the event wasn?t fired because there?s a separate bool property which controls that.