Steve's completions after raise/except (bugs 637/226 - a feature so nice, it's been reported twice).
I've written a patch for this, but the simplest/quickest way I can see to restrict the completions is simply to filter to classes with "Error" or "Exception" in the name (and add StopIteration and GeneratorExit manually). Base class information doesn't seem to exist or be easy to get at through the analysis--is there a way to get at it without hurting the response time too much?

In terms of behaviour, completions appear after typing "raise " and "except " or Ctrl+Space immediately after the statement, but if you've started typing a name/open-paren then it reverts to normal completions