Django context locals:  Locals should list Django context variables  Evaluating in a Django frame should evaluate against the context variables  Adds basic test cases for locals in DjangoDebuggingTests
Adds some test cases for existing functionality:
 New django project
 Add app to a django project

Also:
New Django Projects now get a Templates folder
 Don't stop on TextNode's when a breakpoint is set in a template
Fix automation where we are indexable by the caption instead of the filename (this breaks __init__.py's).
Sets the icon on a .htm/.html files in a Django project to a custom icon (because we'll open it with the template editor after the previous change).
 Currently this icon is the html icon with an X through it, so we'll be forced to replace it.
 Some refactoring to accommodate django debugging tests