Adds pure Python script for FastCGI <-> WSGI gateway, in particular to support Django.
Adds a getting started guide on how to setup the FastCGI script. This isn't fully complete, mainly because we need to figure out exactly how this should be setup. But it covers the basics of getting going. This is based upon the IIS/PHP setup instructions available here: http://learn.iis.net/page.aspx/246/using-fastcgi-to-host-php-applications-on-iis/