Adds syntax color highlighting for Django variable templates.
Updates our projection scheme so our core buffer is typeless and we project that into an elision buffer of the real content type. This was necessary because w/o it the HTML classifier can end up classifying our Django templates and we get highlighting that we don't expect.

Implements the variable parsing using a reg ex and adds an API to produce information about the contained Django variable. Adds unit tests for the parsing.

Updates our TemplateClassifier so that it will use the new variable parsing code to parse variable nodes in the template.

Adds a classification type for the Django template tags and adds the exports for it.

Updates DjangoEditingTests to expect the new tags.


Also:
 Implements GetItems() on DictionaryInfo so we can start inspecting Dictionary objects for variables passed in and tracks the variables passed in our project (still more work to do to wire this all up, but it's a good checkpoint w/ the highlight working).