Changes how we model dictionaries, and some small tweaks to Django intellisense so that we now get completions on filter tags based upon how the templates were rendered.
Django code will frequently contain something like:

from django.template.loader import render_to_string

render_to_string('polls/index.html', {'latest_poll_list': 'foo', 'poll' : 100})

This code causes a template to be rendered with the specified arguments in the dictionary ? that is you can use those variable names in expression blocks and have them get outputted, access members, or apply filters to them. To accommodate this we need to track the strings through the dictionaries and not merge the value types back together. That will let us provide completions based upon what the value is for a specific key.

To accomplish this a new DependentKeyValue class is added which is like a VariableDef but it tracks a key type + a set of value types per key-type. Our DictionaryInfo is then updated to use this DependentKeyValue to hold onto its key/value pairs. Then when we do an index into the dictionary we can return only the matching types based upon the key. If there are no matching types we currently return the union of the types which matches our previous behavior. That's not quite perfect, but it's probably as good as it's going to get for our 1.5 release.

Other small tweaks:
 AsciiString grows a GetHashCode/Equals implementation. This was basically missing and it meant we would create multiple constant objects for each literal string in the code (a, b =