Remove link to IronPython web site in license Remove .djt options in installer Fix classification issue when we have 2 multi-line strings on the same line
699	Resolve import box doesn't show when I hover the mouse on functions.

Fix issue where quick info tag wasn?t showing properly because we were timing out on a search and not finding any items.
	A previous fix for Dev11 was checking to see if the session was dismissed.  But if we don?t add any items then the session is dismissed.

The dev11 change was necessary because the session nulls out the properties when the session dismissed.  So now we pass our SmartTagController in via properties and update the controller directly before the session gets dismissed.  Also fixing an issue where the last char of an identifier isn?t showing the smart tag if it?s followed by a non-identifier char such as comma.

702 - VS startup crash
	We shouldn?t enque the analysis unit until after we?ve finished doing the overview walk, otherwise we can end up processing the unit before the overview walk has actually been done.