714 VS crashes when open project 	Handle more exception types when analyzing search paths
713 Import Intellisense should work after typing ","
	Fix intellisense after both comma space and after Ctrl-Space
	Adds test case to CompletionContextTests.Scenario_MemberCompletion_Imports

712 IronPython REPL runs in __builtin__ and not __main__
	Sets the newly created scopes name to __main__ in the IronPython code path
	Adds test case to ReplWindowTests.IronPythonReplTests.IronPythonModuleName


MPFProj bug #12178 http://mpfproj10.codeplex.com/workitem/12178
	Fixes the name we return from get_CanonicalName so that we are more likely to work w/ the Dev11 designer.

Fixes our ReferenceContainerNode so that it properly fires the event when a reference is added.  Also pulls 2 fields off of every HierarchyNode which is kind of a nice little bonus if you?re counting bytes of overhead per file.

Updates our WPF project template so that it includes references to PresentationCore, PresentationFramework, and WindowsBase.