Suppresses the launching of the ASP.NET Dev server when running under the debugger. When the project is loaded we disable the property on the ASP.NET project system.
- qualified MySQLdb's Python version
- changed ..." Core "... to ..." (Core) "... in the WA Python SDK's title

Fixes running in IIS Express - we now look for the APPCMD env var and use it to try and find the IIS config file.
Set the EMULATED env var in VS like we do in the command line

Few small cleanups in the analysis engine which don't actually change anything, add a commented out test case for something we can't rix right now.