744 [Python Tools] Doesn't properly notify Visual Studio about creating a new folder    Makes the appropriate call after the folder is added to the project
Also adds some stand alone tests for the wfastcgi script.  These are in a new DLL (which is in the wrong place, I'll move it on the next checkin).  This is the start of creating something standalone from AnalysisTest - so far there are just some simple test cases + a stress test case.