This came up from investigating bug 759 (drag drop folder into solution explorer). While I cannot repro the reported issue, I did notice another issue - when we drag and drop a folder containing sub-folders we end up copying all of the items to the same level in the directory structure, but we add the files correctly within solution explorer. Ultimately this means the project is out of sync w/ the project system. Also dragging and dropping to the project node doesn't work correctly.
So now we do the right thing to get the full paths.

Adds a test case which checks both project + folder nodes, the test case uses copy and paste instead of drag and drop because drag and drop is hard to setup in a test. But we go through the same code path in PTVS either way.