Adds new Dev11icons, moves our icons into one central location
To support multiple versions of the icons in .resx files we get a new ?ConditionalEmbeddedResource? item group which our ProjectBefore.settings includes a target for which pre-processes these and creates a normal EmbeddedResource item group.

Moves to using package ID + resource ID for doing resources for our templates

For random icons we refer to the item from our project files using the VSTarget property

For VSCT files we add a VSCTDefinitions which defines DEV11 so we can do conditionals in our .vsct files