821 new project ui tests failures 	Makes some changes to how we find the templates on Dev11 	Changes how we focus the item so that we use the SelectionItemPattern instead of setting focus which no longer works on Dev11 for some reason
820 Set user agent for Azure
	Adds user agent header for all requests

835 Cannot see which file is startup file (VS 2012)
We now bold the startup item, including jumping through some hoops to bold it when the project is initially opened

814 Comment selection comments one extra line
Our mocks were broken here, so our test was passing when it should have failed.
In general in VS the selection is active, but sometimes empty. Our selection was not active very often. So now
it defaults to being active.

Also, a previous fix to MockTextEdit wasn't quite right, we need to have some start position for the deletion. Now both the
comment and method extraction tests pass.


824 Collapse to Definitions not working

Our outlining tags are incorrectly saying they're not implementations, so they never get collapsed


Also fixes ImportWizard so it builds w/ only Dev11 installed