Updates wfastcgi.py for Anteres.  This updates us so that we work both w/ Anteres where we'll be using URL rewriting while at the same time supporting us running w/o Anteres.
Use DOCUMENT_ROOT to find the root of the site - under Anteres shared hosting we need to use this instead of web.config.  This enables us to use 1 shared copy of wfastcgi.py on the system which can then be used for all the different site's.

We also now wait for the 1st request to read this information as before the 1st request we don't have anywhere t o look.

Finally tweaking the way we deal w/ SCRIPT_NAME and PATH_INFO again, and now we basically just rely on PATH_INFO all the time which works across a multiple Python frameworks (Pylons was breaking w/ our old scheme).