Special stackless feature support.
Stackless python can crash on attach since the CFrame object it can inject
into the callchain does not have all the traditional Frame object members
like trace for example.

The debugger and profiler also need to know about how to turn on tracing for
tasklets in stackless python. By default you need to set the trace on every
tasklet that is created, but there is a special flag on stackless that allows
you to turn on global tracing.

The debugger was also having problems stepping over functions that would create
tasklets or cause a context switch in stackless.
So when a function call would cause a context switch, the next line you would
see would be the first line of some random tasklet that the stackless scheduler
chose to run, so when a context switch occurs we turn of the stepping so the
tasklet can run and so we can also hit breakpoints within any running tasklet.

You can also use the context switch to attach the trace to the tasklets, but
since the global trace is set, you don't need to worry about it.