Fixes a couple of issues w/ code formatting as discovered by Hughes' testing.
The line preserving code replacing wasn't handling when the number of lines was reduced, fixed and added a test case
We were also losing comments in places where we were replacing the formatting. We now don't replace the formatting if it contains non-whitespace characters
Finally, there are a few tweaks to the comment reflowing so that we preserve existing new lines rather than flowing the text together. In the urls.py case this prevents us from moving the import statements which are commented out up to the previous line.