Use platform module and its methods whenever available to build implementation/platform description and include bitness in it, for more descriptive and better formatted strings (e.g. was cpython 3.3.0 (win32), now CPython 3.3.0 (Windows 32-bit)).
Made attach server thread a daemon thread, so that it doesn't keep the process alive if main thread has completed execution.

Cleaned up the code that registers trace functions, and made it work for IronPython with tracing & frames enabled.

Added a check for IronPython that bails out early with a descriptive error message if tracing or frames are not enabled.

If remote socket reports connection being dropped, treat it as process exit in the debugger, so that it does not hang if remote process is forcibly terminated before it can report completion.

Removed a bunch of redundant catch(SocketException) in those places where only NetworkStream operations occur (those wrap all SocketExceptions into IOExceptions).