998 Format selection failures when editing
Provides a more consistent model for white space ? removes trailing new line white space on statements, and adds SetLeadingWhiteSpace API which mirrors the existing GetLeadingWhiteSpace API.

Removes the trailing white space after the colon in a suite statement and keeps it on the 1st statement in the suite.

These changes enable cleaning up the SelectionTarget API so that it better serves the dual purpose of working with both extract method and format selection. Now it returns a suite of statements that don't include a colon. Instead ExtractedMethodCreator will create a SuiteStatement which includes a colon when it needs one.

Also fixes up our insertions of new lines when doing a method extraction so we use \n if that's what the buffer is set to use.

Makes the selection when doing reformat selection be the full expression which is getting reformatted so at the end the user has some visual indication of what was reformatted if we expanded the selection.

 Fixes a bug in the line preserving code replacer so that we don't accidently delete a bunch of code

1057 IPython and inline plots are not compatible
 Encode to ASCII so that we can successfully do the base64 decode