1007 from...import completions should not keep appearing when the user has typed 'as' 1035 Intellisense shows classes when it shouldn't 1036 Intellisense shows * when it shouldn't Fixes FromImportCompletionAnalysis to use the correct analysis for the provided tokens. Now ensures the syntax of the statement is correct before providing completions.
1052 Completion filtering should compare against display text rather than insertion text
Fixes references to InsertionText in FuzzyCompletionSet

1058 Automatic method population fails if a keyword argument defaults to '\n'
Fixes FunctionInfo.GetConstantRepr to correctly handle escape characters in strings.

1005 Background analyzer should use correct version for parsing files
Fixes hardcoded version reference in PyLibAnalyzer

1015 Complete word does not complete if filtering is disabled
Adds FuzzyCompletionSet.SelectSingleBest() and updates TriggerCompletionSession in IntellisenseController and DjangoIntellisenseController to use it when attempting to complete a word. SelectSingleBest() ignores the user's filtering preference and returns True if the selected item should be committed immediately.

Also changes the default encoding for Python 3.x to UTF-8 when analysing files. This does not affect the encoding used when opening a file in the IDE.