1597 Clean existing database before reanalyzing PyLibAnalyzer now determines all the files to update before starting and deletes any filegroup that will be updated. Adds database version checks. Removes '/all' switch from 'Refresh DB' button in the environment list. Detection of invalid or partially-generated databases should be reliable enough that the quick refresh will always produce a valid database. Adds tests to check the analyzer is updating the correct files in the database.
1555 Database is not always loaded after reanalysis
Fixes race condition on NotifyNewDatabase() that was due to two separate threads detecting that analysis had completed. Now the status updater thread will only send the notification if we know that analysis was started from a different instance of VS.

Fixes unhandled OOM when initializing analyzer status updaters. The constructed object is now disposed immediately and the exception is saved to be returned in ThrowPendingExceptions(). Since this convention is used for exceptions within the worker thread, it makes sense to also use it for initializing the worker thread and ensure the constructor never throws.
Replaces GetPythonToolsInstallPath() with a static PythonToolsInstallPath.GetFile() that is shared by source code. By looking for a specific file, we can ensure that a location in the GAC is not returned when it would be inappropriate.
Fixes some events that were never disconnected.
Fixes test assembly name in strong name verification scripts.