#1654 No tooltip while editing
Always returning TIP_S_NODEFAULTTIP suppresses the regular quick info. Instead, call IVsDebugger.GetDataTipValue, which will return a proper HRESULT depending on whether debugging is active or not, and whether the expression could be evaluated. This ensures that we always fall back to our usual analysis-driven quick tip provider when data tips are not available.

In addition, this fixes the problem where data tips did not work when hovering over the first character of the identifier.